split hook from global stats
